[Feature] 파일 다운로드 기능 추가

+++ module-infra/src/main/java/com/back2basics/adapter/persistence/post/file/FileEntityRepository.java
@@ -3,7 +3,7 @@ package com.back2basics.adapter.persistence.post.file;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileEntityRepository extends JpaRepository<FileEntity, Integer> {
+public interface FileEntityRepository extends JpaRepository<FileEntity, Long> {
 
     List<FileEntity> findAllByPostId(Long postId);
